Revert part of the changes
