+++ experimental/dds/tree2/src/core/index.ts
@@ -82,8 +82,7 @@ export {
 	ProtoNodes,
 	CursorMarker,
 	isCursor,
-	TreeStatus,
-	treeStatusFromPath,
+	getDetachedFieldContainingPath,
 } from "./tree";
 
 export {
