+++ experimental/dds/tree2/src/feature-libraries/editable-tree-2/editableTreeTypes.ts
@@ -4,7 +4,7 @@
  */
 
 import * as SchemaAware from "../schema-aware";
-import { FieldKey, TreeSchemaIdentifier, TreeValue, TreeStatus } from "../../core";
+import { FieldKey, TreeSchemaIdentifier, TreeValue } from "../../core";
 import { Assume, RestrictiveReadonlyRecord, _InlineTrick } from "../../util";
 import { LocalNodeKey } from "../node-key";
 import {
@@ -19,6 +19,7 @@ import {
 } from "../typed-schema";
 import { EditableTreeEvents } from "../untypedTree";
 import { FieldKindTypes, FieldKinds } from "../default-field-kinds";
+import { TreeStatus } from "../editable-tree";
 import { TreeContext } from "./context";
 
 /**
