+++ experimental/dds/tree2/src/feature-libraries/editable-tree/editableTree.ts
@@ -20,8 +20,6 @@ import {
 	AnchorNode,
 	inCursorField,
 	rootFieldKey,
-	TreeStatus,
-	treeStatusFromPath,
 } from "../../core";
 import { brand, fail } from "../../util";
 import { FieldKind } from "../modular-schema";
@@ -43,7 +41,12 @@ import {
 	contextSymbol,
 	treeStatus,
 } from "../untypedTree";
-import { AdaptingProxyHandler, adaptWithProxy, getStableNodeKey } from "./utilities";
+import {
+	AdaptingProxyHandler,
+	adaptWithProxy,
+	getStableNodeKey,
+	treeStatusFromPath,
+} from "./utilities";
 import { ProxyContext } from "./editableTreeContext";
 import {
 	EditableField,
@@ -52,6 +55,7 @@ import {
 	proxyTargetSymbol,
 	localNodeKeySymbol,
 	setField,
+	TreeStatus,
 } from "./editableTreeTypes";
 import { makeField, unwrappedField } from "./editableField";
 import { ProxyTarget } from "./ProxyTarget";
