+++ experimental/dds/tree2/src/feature-libraries/editable-tree/index.ts
@@ -13,6 +13,7 @@ export {
 	areCursors,
 	localNodeKeySymbol,
 	setField,
+	TreeStatus,
 } from "./editableTreeTypes";
 
 export { isEditableField } from "./editableField";
@@ -56,4 +57,4 @@ export {
 
 export { EditableTreeContext, getEditableTreeContext } from "./editableTreeContext";
 
-export { isPrimitive } from "./utilities";
+export { isPrimitive, treeStatusFromPath } from "./utilities";
