+++ experimental/dds/tree2/src/feature-libraries/index.ts
@@ -51,6 +51,7 @@ export {
 	comparePipeline,
 	compileSyntaxTree,
 	setField,
+	TreeStatus,
 } from "./editable-tree";
 
 export {
